[Extensions] Update worker browser state on worker OnStopped().
Before this change a worker could fail to terminate (i.e.
SWTQ::DidStopServiceWorkerContext() will never be called).
SWTQ::OnStopped() is called in this case, but it only updates
WorkerIdSet tracking of the running extension worker.

After this change we update the browser state back to kInitial in
OnStopped(), but bail out early if the extension is already deactivated.

This will ensure that WorkerState::ready() is more accurate when
dispatching tasks to an extension worker.

Bug: 40936639
Change-Id: I4a57584d2f3106fc700e339ba2e520acf66dbbc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5585866
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348904}