Fix Crash when trying to access ToastController
Chrome was crashing on non-normal browser windows because we were trying
to access the ToastController through a null ToastService. ToastService
is not being initialized for non-normal browser windows because toasts
are not supported in those cases.

This CL adds if guards to existing use cases for trying to access the
ToastController and attempting to show a toast. Documentation is also
added so it's clear to callers that the ToastController can be null.

Fixed: 362846101
Change-Id: I6de5afd6b3e4df1bf23dee705b60688a9c041aea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5826865
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348970}