HttpStreamPool: Add methods to get the current information
Some tests in NetworkContextTest use
ClientSocketPool::GetInfoAsValue() to inspect the current information
of the pool. Add similar methods to the HttpStreamPool so that we can
use these methods to fix NetworkContextTests when the HttpStreamPool
is used. We could also use these methods to export the information
to the NetLog in the future.

Bug: 346835898
Change-Id: I285dc70551a16da965ec2fd342147f7fb75ccd48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5824033
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348971}