[Text Contrast] Remove static variable for feature flag
I noticed during local development that the test I added recently
(FontRenderParamsTest::IncreasedContrast) always fails on the first
run. This is because I store the value of the feature flag in a static
variable, so the ScopedFeatureList doesn't apply and instead uses
the previously queried value.

It passes on rerun because it is run in isolation, so the static
variable gets re-initialized.

The approach here is to remove the static caching of the feature flag
value.

Bug: 362788342
Change-Id: I2e1080c40bc0bab541474250210a12331671af2e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5825114
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1348992}