Fix plumbing for HistoryClustersHandler
This CL fixes several edge cases. There are no intended user-visible
behavior changes for the common use cases.

Prior to this CL, HistoryClustersHandler used chrome::FindLastActive()
and chrome::FindTabbedBrowser() to get a Browser*. There is no guarantee
that these functions will return the correct browser. This CL fixes the
plumbing to pass through a BrowserWindowInterface from the constructor.

The test suite HistoryClustersHandlerTest uses BrowserWithTestWindowTest
which is a known bad-pattern. The test
QueryClustersResultToMojom_Integration is modified to become a
browser_test. The test OpenVisitUrlsInTabGroup is a duplicate of an
existing browser_test and is deleted.

Change-Id: Ia832c826b8093fb24ee2a3e4f00c5035bcb27315
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5809336
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Reviewed-by: Marlon Facey <mfacey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349017}