Add kBoostRenderProcessForLoadingPrioritizePrerenderingOnly feature param
We came up with an assumption that prioritizing the renderer process
that hosts the prerendering page could improve performance because we
often see that prerendered page immediately activated after prerendering
is started.

This CL adds kBoostRenderProcessForLoadingPrioritizePrerenderingOnly
feature param to check if the above theory is correct or not.

Relevant: PrerenderBrowserTest.VerifyPrerenderProcessVisibility

Bug: 351953350
Bug: 40183020
Change-Id: Iaad4aa9d6c3572e77704260757564f711e548888
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5821335
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349047}