PA: Upgrade `DCheckIsValidSlotSpan()`
This means adding more `CHECK()`s to the shipping build, asserting that
PA is dispensing valid slot spans. A few cycles on competitive
benchmarks suggest that we can get away with this (perf-wise) as long as
we don't upgrade the `DCHECK()` in `partition_page.h` to a full-on
`CHECK()`.

Fixed: 40796496
Change-Id: Iaceee21361a531948814618fc225cedad4872439
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5771956
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Commit-Queue: Keishi Hattori <keishi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349089}