Relink GSA and google_bottom_bar
This CL is kind of revert the https://crrev.com/c/5817874.
We will delete all the GSA library for icing, but still want to centralize the GSA related functionalites, so GSAUtils are created
in https://crrev.com/c/5827164, and we will use GSAUtils then.

Bug: b/362498780

Change-Id: I522704a7b16ebd316c41baa4fd538ce9fc83b3ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5827877
Reviewed-by: Edmund Wright <edmundw@google.com>
Commit-Queue: Edmund Wright <edmundw@google.com>
Cr-Commit-Position: refs/heads/main@{#1349103}