Allow registering KeyedServiceFactory while no context exists
When KeyedServiceFactory infrastructure is compiled with tight
registration, it is known that all context will be marked live
when created and marked dead when destroyed.

Thus in that case use a counter to check whether there are any
live context when a KeyedServiceFactory is registered. This
allow tests to register KeyedServiceFactory dynamically as long
as they have not created any context yet.

This loosen the restriction a bit, but is safe since for prod,
DisallowKeyedServiceFactoryRegistration(...) is called before
the first context is created.

Bug: none
Change-Id: I54017e75e5d8923507f441336241e29891acf27a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5822768
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349106}