[Settings] Isolate wide screen padding logic from SettingsActivity
SettingsActivity has multiple functionalities, and one of them is to
apply padding to its fragment for wide screens. This patch extracts the
functionality to an isolated class to make the upcoming refactoring of
SettingsActivity easier.

Bug: b/356743945
Change-Id: Ib13c598bb23bd7b5a41d704ac3e9716240e1072f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5820236
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349144}