[iOS] Add notification opt-in button for Magic Stack modules
This CL introduces a generic mechanism to add a "Turn On Notifications"
button to Magic Stack modules. This allows users to opt-in to push
notifications for any module that supports this functionality. The
Safety Check Notifications project will be the first to utilize this new
Magic Stack action. This change is part of the broader Safety Check
Notifications project.

Change-Id: I267eed2aff24ea5077924aed74ca9c436dae3c1c
Fixed: 363098810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5826320
Reviewed-by: Scott Yoder <scottyoder@google.com>
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1349159}