Add a refresh button for inline plus address suggestions.
This CL adds a factory function for creating inline plus suggestions
in the Autofill popup. This CL focuses primarily on wiring up the
action for clicking on the refresh button and leaves various TODOs to
fix the actual style of the suggestion.

Bug: 362445807
Change-Id: I2007e35707baa0195390c8daf18bc00fdfe54118
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5816596
Reviewed-by: Bruno Braga <brunobraga@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1349171}