ios: Deflake PasteboardUtilTest.StoreInPasteboardWorks.
BrowserViewControllerTest
presentIncognitoReauthDismissesPresentedState calls
-presentViewController on a detached view controller (which in and
of itself will soon be a fatal error) and leaks a BVC,leading to all
sorts of unexpected results.

The specific error was:
Presenting view controller <UIViewController: 0x112010bc800> from
detached view controller <BrowserViewController: 0x112014f1080> is not
supported, and may result in incorrect safe area insets and a corrupt
root presentation. Make sure <BrowserViewController: 0x112014f1080> is
in the view controller hierarchy before presenting from it. Will become
a hard exception in a future release.

The flake from this leak was due to the OmniboxViewController holding
on to a UIPasteboardChangedNotification, which expected a
ClipboardRecentContent::GetInstance. PasteboardUtilTest does not and
should not have to set this.

Making the BVC a rootViewController avoids the
presentation error and avoids leaking the view, which fixes the
StoreInPasteboardWorks flake.

Change-Id: I91526c19379f568741d2fed0fa955ac093718f16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823836
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349198}