[Add User after Enrollment] Launch the login on feature flag.
This cl is part of the go/add-user-during-enrollment.
Context: the signin artifacts and the refresh token are stored in the
wizard context. This cl skips the second gaia sign in, by launching
the login flow based on that stored information if the feature flag
for this feature is set.

Change-Id: I280508df8225b0263570c42d93327d48b580ea5f
Bug: b:361289699
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5803241
Reviewed-by: Renato Silva <rrsilva@google.com>
Reviewed-by: Andrey Davydov <andreydav@google.com>
Reviewed-by: Leon Masopust <lmasopust@google.com>
Commit-Queue: Stanisław Solarewicz <ssolarewicz@google.com>
Cr-Commit-Position: refs/heads/main@{#1349213}