[iOS] Add Safety Check notifications toggle on module long press
This CL adds the ability to manage Safety Check notifications directly
from the Safety Check card in the Magic Stack. A long-press on the card
now presents an option to toggle Safety Check notifications on or off,
based on the current notification opt-in state.

Change-Id: I73ee62d47b73059b61b7a9105cb5c1a5c0f8d1bf
Fixed: 363212339
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5826322
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1349225}