Return early in IsClipboardCopyAllowedByPolicy on empty endpoint
The DCHECKs in that function are wrong as
`source.data_transfer_endpoint()` can be null, which leads to a crash in
later code. An early return is added to avoid that issue until the root
cause of the issue is fully understood.

Bug: 362674391
Change-Id: I2fe3fc470061d881a958503907c37a0ca075d0d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5827397
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Reviewed-by: Nancy Xiao <nancylanxiao@google.com>
Commit-Queue: Nancy Xiao <nancylanxiao@google.com>
Auto-Submit: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349244}