[iOS][LO] Infill cropped screenshot insets for Lens Overlay
This CLs adds an infill on top and bottom, corresponding to the insets
that were cropped as a result of capturing only the content area.

The side effect of this Cl is that the animation is much smoother as the
screenshot, now fulscreen, overlaps the content area behind perfectly,
making for a natural transition between the 2 states.

Fixed: 359781942
Change-Id: I021d299abcfe03ec86674456e5b6e832d282043b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823627
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Christian Xu <christianxu@chromium.org>
Commit-Queue: Christian Xu <christianxu@chromium.org>
Auto-Submit: Radu Nitescu <radunitescu@google.com>
Reviewed-by: Federica Germinario <fedegermi@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349262}