[UPMAccessLossWarning] Shut down Chrome after redirecting user to GMS
After the export flow Chrome needs to be restarted to update the user
state:
- If no GMS Core is installed, we need to restart Chrome so that
kEmptyProfileStoreLoginDatabase would be set and password saving would
be disabled.
- If there was a migration error and the user resolved it by exporting
and importing their passwords manually to GMS Core, Chrome needs to be
shut down (not restarted to avoid conflict when redirecting to GMS Core)
to set UseUpmLocalAndSeparateStoresState to On state on startup.

from tests can't be tested.

Bug: 354876096
Change-Id: I54d7eb5e9a1b2b16079dd5dbd8a7f07916a85904
Low-Coverage-Reason: The changes in the constructor, which is not called
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5827532
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Cr-Commit-Position: refs/heads/main@{#1349320}