Only emit chrometto traces for main overview mode cases.
Primarily, we don't want to capture traces from starting overview
for "informed restore" after login. The first round of traces had
mostly these cases, and it's time consuming to sift through them.

The trace now only starts when either the user hits the
overview button in clamshell mode or drags a window from shelf
in tablet mode. These are the 2 main use cases in the field and
ones we're focused on currently for performance.

Bug: b:362822878
Change-Id: I5de75746c318c7a37437fdc501be795877553690
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5828042
Commit-Queue: Eric Sum <esum@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349330}