WebUI: Add a Lit version of FocusRowMixin
This is in preparation for migrating downloads-item, which uses this
mixin, to Lit. While originally intended for iron-list children, this
mixin's functionality is not specific to iron-list, and also applies
to children of any parent list that provides a tabindex for a template
and focuses items on click. The Lit-based infinite-list in history
clusters, which is planned for re-use by the downloads UI, also has
the same behavior. Moreover, Lit-based elements are sometimes used as
iron-list children during migrations.

Bug: 40943652
Change-Id: Ia2b0b3c2ac88940025ca2d03a585911be4efb713
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5826046
Reviewed-by: John Lee <johntlee@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349337}