Move haptics feedback to the AppListItemView
For drag and drpo refactoring, the AppsGridView is less knowledgeable of
the whole drag and drop sequence. While haptics events are supposed to
be sent only once at the beginning of the drag operation, the
AppsGridView would take action every time a view is dragged within it.
This means that if the view is dragged in and out and in again of the
AppsGridView, the view considers two different drag events even if the
drag was not released.

To ease that, move the haptics implementation out of the AppsGridView
and into the AppListItemView when it is set to MouseDragging. This way
we can guarantee that the haptics events in only sent once per drag
operation

Also, fix an issue where the drag pointer was not correctly recognized
by the AppsGridView when using the DragAndDropController. Since
DragEvents are not considered either MouseEvent or TouchEvent, but its
own type, using IsMouseEvent() would always return false for this event.
Instead, expose the DragAndDropController from Shell and fetch the event
source from there.

Bug: None
Change-Id: Ia4237ac6650a5bde72b3709e5628d06ebaf7b269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5824332
Reviewed-by: Toni Barzic <tbarzic@google.com>
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349350}