[Extensions] Create a common base class for two somewhat-related APIs.
The bookmarks and bookmarkManagerPrivate API are related, but the
current implementation ties the latter to the former through a
shared base class for the function implementations.

This CL splits that base class out into a separate implementation
that lives in chrome/browser/extensions/api/bookmarks_core. This
breaks the dependency between the two.

This is primarily a cut-and-paste change, with some modernization
of braces, etc. It also cleans up some unused includes.

Bug: 40593486
Change-Id: I4d181049f95b2c1637db3737d5592c707b5fe9bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5827167
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Auto-Submit: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349364}