Make KeyboardFocusableScrollers keyboard-focus but not mouse-focus
This (re-)introduces the concept of IsMouseFocusable, which
represents whether an element should be focused by the mouse. This
is now distinct from IsKeyboardFocusable: elements can be one of
these without being the other. Both require IsFocusable() to be
true.

This behavior is guarded behind the flag KeyboardFocusableScrollers.
When the flag disabled, IsMouseFocusable and IsKeyboardFocusable will
always return the same value. With the flag enabled, scrollers
(without an explicit tabindex) will be keyboard focusable, but will
not be mouse focusable.

Along the way, I ended up (thanks to a comment from dbaron@) doing
a refactor of `SupportsFocus()` and `IsFocusable()` (now called
`IsFocusableState()`) to return a tri-state enum, which explicitly
keeps track of whether the element supports or is focusable due to
keyboard focusable scrollers. That ends up making the logic for
`IsMouseFocusable()` and `IsKeyboardFocusable()` a lot easier to
grok since there aren't hidden dependencies on the logic in
`SupportsFocus()`. I also removed the default parameter values for
both methods, since they were passed explicitly almost everywhere.

Bug: 361072782
Co-authored-by: Di Zhang <dizhangg@chromium.org>
Change-Id: I4f6e2a4f4e6b27ec637315ea17aafe6da39ef235
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5812197
Reviewed-by: David Baron <dbaron@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349407}