Ensure CCT Tabs gets TabUma metrics
Seems like it was most likely an accident of history that CCT tabs
don't get TabUma metrics (including when you open a CCT Tab in Chrome).

This was discovered because the CCTPrewarmTab experiment accidentally
caused TabUma TabHelpers to be attached to CCT tabs and metrics counts
for tabs becoming foregrounded shot up.

So instead of intentionally excluding CCT, or only getting metrics when
a Tab is pre-warmed, make sure a TabUma helper is always created for
CCT.

Bug: 40286920
Change-Id: I117cea9c346d2b5655f12a8cf6538dc905a0622f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823287
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349419}