Disable Canvas2dTextMetricsShaping
Direct text shaping when `measureText()` is called was implemented on
crrev.com/c/5758312. That CL included a flag to be used as kill switch
if a performance regression was reported.

crbug.com/362784008 reported a regression on Speedometer scores after
the change. This CL disables said flag as a result.

Bug: 362784008, 341213359
Change-Id: I7f27dfb356a57095e970ba2a8bb2992cd58a34a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5829243
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349451}