Also consider stretch/-webkit-fill-available in code that used to use Length::IsSpecified.
This fixes an error in commit 57b34a9c9d1c20f24307766f3c16242c3a6365e8,
which converted callers of Length::IsSpecified to handle calc-size().
The changes in that CL didn't consider the -webkit-fill-available (or,
in the future, stretch) keyword, which is not covered by
HasAutoOrContentOrIntrinsic().

This makes these callers consider that case as well.

Fixed: 361601751
Bug: 313072
Change-Id: I9ecaf6b6f3b0c108e33068325b302a5100f1fb28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5823770
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1349460}